Move replacements from MainActivity to ReplacementsFragment
