+++ app/src/main/java/com/example/planlekcji/MainActivity.java
@@ -2,16 +2,11 @@ package com.example.planlekcji;
 
 import android.content.Context;
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.content.pm.ActivityInfo;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
 import android.os.Bundle;
-import android.text.Html;
-import android.view.View;
-import android.widget.EditText;
 import android.widget.ImageButton;
-import android.widget.TextView;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.app.AppCompatDelegate;
@@ -19,21 +14,16 @@ import androidx.viewpager2.widget.ViewPager2;
 
 import com.example.planlekcji.fragments.ViewPagerAdapter;
 import com.example.planlekcji.settings.ui.SettingsActivity;
-import com.example.planlekcji.utils.BoyerMooreSearch;
-import com.example.planlekcji.utils.DelayedSearchTextWatcher;
 import com.example.planlekcji.utils.ToastUtils;
+
 import com.google.android.material.tabs.TabLayout;
 import com.google.android.material.tabs.TabLayoutMediator;
 
-import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 public class MainActivity extends AppCompatActivity {
 
     private static Context appContext;
 
-    private List<String> replacements;
     private ViewPager2 viewPager2_appContent;
 
     @Override
@@ -82,23 +72,7 @@ public class MainActivity extends AppCompatActivity {
                     break;
             }
         }).attach();
-
     }
-
-//    @Override
-//    protected void onStop() {
-//        super.onStop();
-//        EditText searchBar = findViewById(R.id.editText_searchBar);
-//
-//        SharedPreferences sharedPref = this.getSharedPreferences("sharedPrefs", 0);
-//        SharedPreferences.Editor editor = sharedPref.edit();
-//        editor.putString(getString(R.string.searchKey), String.valueOf(searchBar.getText()));
-//        editor.apply();
-//
-//        TabLayout tabLayout = findViewById(R.id.tabLayout_timetableDays);
-//        mainViewModel.setSelectedTabNumber(tabLayout.getSelectedTabPosition() + 1);
-//    }
-
 
     private void setEventListenerToSettingsButton() {
         ImageButton button = findViewById(R.id.imageButton_goSettings);
@@ -108,91 +82,8 @@ public class MainActivity extends AppCompatActivity {
         });
     }
 
-    private void setEventListenerToSearchBar() {
-        EditText searchBar = findViewById(R.id.editText_searchBar);
-        TextView textView_noResults = findViewById(R.id.textView_noResults);
-
-        searchBar.addTextChangedListener(new DelayedSearchTextWatcher(query -> {
-            if(query.isEmpty()) {
-                setReplacements();
-            } else {
-                String foundData = searchReplacements(query);
-                if(foundData.isEmpty()) {
-                    textView_noResults.setVisibility(View.VISIBLE);
-                    setReplacements();
-                } else {
-                    textView_noResults.setVisibility(View.GONE);
-                    setReplacements(foundData);
-                }
-            }
-        }));
-    }
-
-    private void setReplacements() {
-        TextView textFieldReplacements = findViewById(R.id.textView_replacements);
-        EditText searchBar = findViewById(R.id.editText_searchBar);
-
-        if(replacements.isEmpty()) {
-            searchBar.setVisibility(View.GONE);
-            textFieldReplacements.setText(getString(R.string.no_replacements));
-        } else {
-            searchBar.setVisibility(View.VISIBLE);
-            textFieldReplacements.setText(Html.fromHtml(String.join("<br><br>", replacements), Html.FROM_HTML_MODE_LEGACY));
-        }
-    }
-    private void setReplacements(String data) {
-        if (data != null) {
-            TextView textFieldReplacements = findViewById(R.id.textView_replacements);
-            textFieldReplacements.setText(Html.fromHtml(data, Html.FROM_HTML_MODE_LEGACY));
-        }
-    }
-
     public static Context getContext() {
         return appContext;
-    }
-
-    private String searchReplacements(String searchingKey) {
-        if(searchingKey.isEmpty()) return "";
-
-        // in case user types '4ptn' instead of '4 ptn'
-        searchingKey = handleOtherUserInput(searchingKey).toLowerCase();
-
-        BoyerMooreSearch boyerMooreSearch = new BoyerMooreSearch();
-
-        // searching
-        StringBuilder foundResults = new StringBuilder();
-        for (int i = 1; i < replacements.size(); i++) {
-            String singleReplacement = replacements.get(i);
-
-            String text = singleReplacement.toLowerCase();
-
-            if(boyerMooreSearch.search(text, searchingKey)) {
-                if(foundResults.length() == 0) foundResults.append(singleReplacement);
-                else foundResults.append("<br><br>").append(singleReplacement);
-            }
-        }
-        return foundResults.toString();
-    }
-
-    private String handleOtherUserInput(String searchingKey) {
-        Pattern pattern = Pattern.compile("\\d");
-        Matcher matcher = pattern.matcher(searchingKey.substring(0, 1));
-
-        if(!searchingKey.contains(" ") && searchingKey.length() > 1 && matcher.find()) {
-            searchingKey = searchingKey.charAt(0)+" "+searchingKey.substring(1);
-        }
-
-        return searchingKey;
-    }
-
-    private String updateSearchBar() {
-        SharedPreferences sharedPref = this.getSharedPreferences("sharedPrefs", 0);
-        String searchKey = sharedPref.getString(getString(R.string.searchKey), "");
-
-        EditText searchBar = findViewById(R.id.editText_searchBar);
-        searchBar.setText(searchKey);
-
-        return searchKey;
     }
 
     public boolean isOnline() {
