+++ app/src/main/res/layout/activity_main.xml
@@ -51,71 +51,6 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <ScrollView
-        android:id="@+id/scrollView_replacements"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginBottom="8dp"
-        android:visibility="invisible"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/tabLayout_navigate">
-
-        <LinearLayout
-            android:id="@+id/linearLayout_replacements"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical">
-
-            <EditText
-                android:id="@+id/editText_searchBar"
-                android:layout_width="match_parent"
-                android:layout_height="48dp"
-                android:layout_marginStart="4dp"
-                android:layout_marginEnd="8dp"
-                android:autofillHints="searchQuery"
-                android:ems="10"
-                android:hint="@string/searchHint_searchBar"
-                android:inputType="textPersonName" />
-
-            <TextView
-                android:id="@+id/textView_noResults"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="8dp"
-                android:layout_marginTop="4dp"
-                android:layout_marginEnd="8dp"
-                android:layout_marginBottom="4dp"
-                android:text="@string/noResults_text"
-                android:textSize="18sp"
-                android:textStyle="bold"
-                android:visibility="invisible" />
-
-            <View
-                android:id="@+id/divider"
-                android:layout_width="match_parent"
-                android:layout_height="1dp"
-                android:layout_marginTop="8dp"
-                android:layout_marginBottom="8dp"
-                android:background="?android:attr/listDivider" />
-
-            <TextView
-                android:id="@+id/textView_replacements"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:layout_marginStart="8dp"
-                android:layout_marginTop="8dp"
-                android:layout_marginEnd="8dp"
-                android:layout_marginBottom="8dp"
-                android:scrollbars="vertical"
-                android:textSize="18sp" />
-        </LinearLayout>
-    </ScrollView>
-
     <ImageButton
         android:id="@+id/imageButton_goSettings"
         android:layout_width="60dp"
